final commit bug fixes 1

+++ flow-of-control/ex1.php
@@ -1,8 +1,8 @@
 <?php declare(strict_types=1);
 
 //Input all three numbers in an array
-$input[0] = readline('Input the 1st number: ');
-$input[1] = readline('Input the 2nd number: ');
-$input[2] = readline('Input the 3rd number: ');
+$input[0] = (int) readline('Input the 1st number: ');
+$input[1] = (int) readline('Input the 2nd number: ');
+$input[2] = (int) readline('Input the 3rd number: ');
 //Output the maximum in of the array
 echo max($input) . PHP_EOL;
